Remember hidden state on startup
One issue that had to be dealt with is that the GtkWindow has to be loaded for
everything to work properly. At the same time, there were issues with hiding
all the notes on startup as they would then default to the 200 x 200 default
size. This seems to be the best solution.